move digest creation to remediation folder
